Enforce strict block signature checks and align signature handling with spec

+++ include/lantern/consensus/signature.h
@@ -14,19 +14,19 @@ void lantern_signature_zero(LanternSignature *signature);
 bool lantern_signature_verify(
     const uint8_t *pubkey_bytes,
     size_t pubkey_len,
-    uint32_t epoch,
+    uint64_t epoch,
     const LanternSignature *signature,
     const uint8_t *message,
     size_t message_len);
 bool lantern_signature_verify_pk(
     const struct PQSignatureSchemePublicKey *pubkey,
-    uint32_t epoch,
+    uint64_t epoch,
     const LanternSignature *signature,
     const uint8_t *message,
     size_t message_len);
 bool lantern_signature_sign(
     const struct PQSignatureSchemeSecretKey *secret_key,
-    uint32_t epoch,
+    uint64_t epoch,
     const uint8_t *message,
     size_t message_len,
     LanternSignature *out_signature);
