+++ src/consensus/signature.c
@@ -33,7 +33,7 @@ void lantern_signature_zero(LanternSignature *signature) {
 bool lantern_signature_verify(
     const uint8_t *pubkey_bytes,
     size_t pubkey_len,
-    uint32_t epoch,
+    uint64_t epoch,
     const LanternSignature *signature,
     const uint8_t *message,
     size_t message_len) {
@@ -52,7 +52,7 @@ bool lantern_signature_verify(
 
 bool lantern_signature_verify_pk(
     const struct PQSignatureSchemePublicKey *pubkey,
-    uint32_t epoch,
+    uint64_t epoch,
     const LanternSignature *signature,
     const uint8_t *message,
     size_t message_len) {
@@ -75,7 +75,7 @@ bool lantern_signature_verify_pk(
 
 bool lantern_signature_sign(
     const struct PQSignatureSchemeSecretKey *secret_key,
-    uint32_t epoch,
+    uint64_t epoch,
     const uint8_t *message,
     size_t message_len,
     LanternSignature *out_signature) {
